+++ node/core/provisioner/src/metrics.rs
@@ -14,7 +14,7 @@
 // You should have received a copy of the GNU General Public License
 // along with Polkadot.  If not, see <http://www.gnu.org/licenses/>.
 
-use crate::disputes::with_staging_api::PartitionedDisputes;
+use crate::disputes::prioritized_selection::PartitionedDisputes;
 use polkadot_node_subsystem_util::metrics::{self, prometheus};
 
 #[derive(Clone)]
